refactoring variable helper
